+++ clare/clare/retry/policy_builder.py
@@ -1,7 +1,140 @@
 # -*- coding: utf-8 -*-
 
+import abc
+
 from . import continue_strategies
 from .policy import Policy
+
+
+class INotifyable(object):
+
+    __metaclass__ = abc.ABCMeta
+
+    @abc.abstractmethod
+    def notify(self, event):
+
+        """
+        Parameters
+        ----------
+        event : clare.retry.policy.IJsonSerializable
+
+        Returns
+        ------
+        None
+        """
+
+        pass
+
+
+class Observable(INotifyable):
+
+    def __init__(self):
+        self._observers = set()
+
+    def register(self, observer):
+
+        """
+        Parameters
+        ----------
+        observer : clare.retry.policy_builder.INotifyable
+        """
+
+        self._observers.add(observer)
+
+    def notify(self, event):
+
+        """
+        Parameters
+        ----------
+        event : clare.retry.policy.IJsonSerializable
+        """
+
+        for observer in self._observers:
+            observer.notify(event=event)
+
+    def __repr__(self):
+        repr_ = '{}()'
+        return repr_.format(self.__class__.__name__)
+
+
+class ObservableFactory(object):
+
+    def build(self):
+
+        """
+        Returns
+        ----------
+        clare.retry.policy_builder.INotifyable
+        """
+
+        observable = Observable()
+        return observable
+
+
+class Broker(object):
+
+    def __init__(self, observable_factory):
+
+        """
+        Parameters
+        ----------
+        observable_factory : clare.retry.policy.ObservableFactory
+        """
+
+        self._observable_factory = observable_factory
+        self._topic_index = dict()
+
+    def create_topic(self, name):
+
+        """
+        A new topic is created only if one by the same name does not
+        already exist.
+
+        Parameters
+        ----------
+        name : str
+        """
+
+        if name not in self._topic_index:
+            self._topic_index[name] = self._observable_factory.build()
+
+    def list_topics(self):
+
+        """
+        Returns
+        -------
+        collections.Sequence
+        """
+
+        return self._topic_index.keys()
+
+    def publish(self, event, topic_name):
+
+        """
+        Parameters
+        ----------
+        event : clare.retry.policy.IJsonSerializable
+        topic_name : str
+        """
+
+        observable = self._topic_index[topic_name]
+        observable.notify(event=event)
+
+    def subscribe(self, subscriber, topic_name):
+
+        """
+        Parameters
+        ----------
+        subscriber : clare.retry.policy_builder.INotifyable
+        topic_name : str
+        """
+
+        observable = self._topic_index[topic_name]
+        observable.register(observer=subscriber)
+
+    def __repr__(self):
+        repr_ = '{}(observable_factory={})'
+        return repr_.format(self.__class__.__name__, self._observable_factory)
 
 
 class PolicyBuilder(object):
